FlexStack/FlexQueue Tests
Summary:
Code refactoring changes made so that things are easier to view. Flex naming structure is more consistent. Requested changes from the audit have been made.

Update: Going along with the changes that have been made, we're back to debugging. This differential does not compile in its current state.

Test Plan: Next differential should feature the newest version of Flex Array and it's corresponding tests (adding more for the removal features). Once that is diffed and reviewed, similar changes will be applied to the rest of the structures ince flex array is the base class for how the rest are structured.

Reviewers: nsgroggett, jcmcdonald

Reviewed By: nsgroggett, jcmcdonald

Subscribers: mrparkman, jenkins, jcmcdonald

Tags: #programming_dept, #bss_team, #pawlib_core_pawlib

Differential Revision: https://phabricator.mousepawmedia.net/D152